# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Urja Meter Ops API",
    "description": "A clean, documented REST API over the legacy Urja Meter Ops portal. Provides programmatic access to smart meter data, distribution transformer info, and energy consumption readings.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/transformers": {
      "get": {
        "summary": "List transformers",
        "tags": [
          "Transformers"
        ],
        "description": "Paginated list of distribution transformers (DTs).",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string",
                            "description": "Transformer code (e.g. DT-001)"
                          },
                          "name": {
                            "type": "string",
                            "description": "Human-readable name (e.g. Malviya Nagar DT 1)"
                          },
                          "feederCode": {
                            "type": "string",
                            "description": "Feeder code this DT belongs to"
                          },
                          "capacityKva": {
                            "type": "integer",
                            "description": "Transformer capacity in kVA"
                          },
                          "syncedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers/{code}": {
      "get": {
        "summary": "Get transformer detail",
        "tags": [
          "Transformers"
        ],
        "description": "Single distribution transformer by code.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "code",
            "required": true,
            "description": "Transformer code (e.g. DT-001)"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Transformer code (e.g. DT-001)"
                        },
                        "name": {
                          "type": "string",
                          "description": "Human-readable name (e.g. Malviya Nagar DT 1)"
                        },
                        "feederCode": {
                          "type": "string",
                          "description": "Feeder code this DT belongs to"
                        },
                        "capacityKva": {
                          "type": "integer",
                          "description": "Transformer capacity in kVA"
                        },
                        "syncedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers/{code}/meters": {
      "get": {
        "summary": "List meters under a transformer",
        "tags": [
          "Transformers"
        ],
        "description": "All smart meters attached to a specific distribution transformer. This is a derived endpoint — the portal doesn't have a direct equivalent.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "code",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "meterId": {
                            "type": "string"
                          },
                          "serialNo": {
                            "type": "string"
                          },
                          "make": {
                            "type": "string"
                          },
                          "phaseType": {
                            "type": "string"
                          },
                          "installStatus": {
                            "type": "string"
                          },
                          "installType": {
                            "type": "string"
                          },
                          "build": {
                            "type": "string"
                          },
                          "dtCode": {
                            "type": "string"
                          },
                          "geo": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "latitude": {
                                "type": "number"
                              },
                              "longitude": {
                                "type": "number"
                              }
                            }
                          },
                          "syncedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "transformerCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
}